Add unit tests for TargetResolveRule
Extracts TargetResolveRule from a private nested record in TargetEditor
to a top-level public record in the same internal package, with a public
static forHandle(ITargetHandle) factory. Adds the pde.ui.tests bundle as
an x-friend of org.eclipse.pde.internal.ui.editor.targetdefinition so
the new TargetResolveRuleTest can verify the rule's contract: same
memento serializes, distinct mementos stay independent, foreign rules
do not conflict, and the CoreException fallback keeps the same handle
instance serialized while distinguishing different broken handles.
